pyproject.toml
[project]
name = "langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.11",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.9",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.2",
    "streamlit>=1.54.0",
    "torch>=2.10.0",
    "transformers>=4.57.6",
]

[dependency-groups]
dev = [
    "langchain",
]

[tool.uv.sources]
langchain = { workspace = true }
